requirements.txt
Pillow>=9.0.0
pytesseract>=0.3.8
PyMuPDF>=1.19.0
tk>=0.1.0